Moved primary driving code into subsystem and fixed bug applying conversion factor twice
In order to use the code for driving the robot in multiple commands and with no interruption in slew rate limiting, I moved the swerve driving code from the TeleOpDrive command to the Drivetrain subsystem.

The bug was that I was multiplying the inputs by the max speed of the motors because that was how the conversion was done last year, but the position/velocity conversion factors for the SparkMAXs already solves that problem, so the multiplication was turning the inputs into garbage.